[FEM] Elmer writer: sort out flow equation
- sort out flow equation - next step to refactor writer.py
